src/pyproject.toml
[tool.poetry]
name = "avito-parser-server"
version = "0.1.0"
description = ""
authors = ["Tim <71645120+TimurStorm@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^21.3.4"
black = "^21.4b0"
python-dotenv = "^0.17.0"
vk-api = "^11.9.1"
aiovk = "^4.0.0"
asyncpg = "^0.22.0"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
